+++ converter/src/main/java/org/geysermc/pack/converter/converter/ExtractionContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2023 GeyserMC. http://geysermc.org
+ * Copyright (c) 2025-2025 GeyserMC. http://geysermc.org
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  *  of this software and associated documentation files (the "Software"), to deal
@@ -26,32 +26,12 @@
 
 package org.geysermc.pack.converter.converter;
 
-import org.geysermc.pack.converter.PackConversionContext;
-import org.geysermc.pack.converter.PackConverter;
-import org.geysermc.pack.converter.data.ConversionData;
+import org.geysermc.pack.bedrock.resource.BedrockResourcePack;
 import org.geysermc.pack.converter.util.LogListener;
-import org.jetbrains.annotations.NotNull;
 import team.unnamed.creative.ResourcePack;
 
-import java.nio.file.Path;
+import java.util.Optional;
 
-public interface Converter<T extends ConversionData> {
-
-    void convert(@NotNull PackConversionContext<T> context) throws Exception;
-
-    T createConversionData(@NotNull ConversionDataCreationContext context);
-
-    default boolean isExperimental() {
-        return false;
-    }
-
-    record ConversionDataCreationContext(
-        @NotNull PackConverter converter,
-        @NotNull LogListener logListener,
-        @NotNull Path inputDirectory,
-        @NotNull Path outputDirectory,
-        @NotNull ResourcePack javaResourcePack,
-        @NotNull ResourcePack vanillaResourcePack
-    ) {
-    }
+public record ExtractionContext(@Deprecated(forRemoval = true) BedrockResourcePack bedrockResourcePack,
+                                Optional<ResourcePack> vanillaPack, LogListener logListener) implements LogHelpers {
 }
