+++ converter/src/main/java/org/geysermc/pack/converter/converter/lang/BedrockLanguage.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025 GeyserMC. http://geysermc.org
+ * Copyright (c) 2025-2025 GeyserMC. http://geysermc.org
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  *  of this software and associated documentation files (the "Software"), to deal
@@ -24,9 +24,9 @@
  *
  */
 
-package org.geysermc.pack.converter.newconverter;
+package org.geysermc.pack.converter.converter.lang;
 
-import org.geysermc.pack.converter.util.LogListener;
+import java.util.Map;
 
-public record CollectionContext(String textureSubDirectory, LogListener logListener) implements LogHelpers {
+public record BedrockLanguage(String language, Map<String, String> strings) {
 }
