+++ converter/src/main/java/org/geysermc/pack/converter/converter/model/ModelConverter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2024 GeyserMC. http://geysermc.org
+ * Copyright (c) 2025-2025 GeyserMC. http://geysermc.org
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  *  of this software and associated documentation files (the "Software"), to deal
@@ -26,7 +26,6 @@
 
 package org.geysermc.pack.converter.converter.model;
 
-import com.google.auto.service.AutoService;
 import net.kyori.adventure.key.Key;
 import org.geysermc.pack.bedrock.resource.BedrockResourcePack;
 import org.geysermc.pack.bedrock.resource.models.entity.ModelEntity;
@@ -41,10 +40,12 @@ import org.geysermc.pack.bedrock.resource.models.entity.modelentity.geometry.bon
 import org.geysermc.pack.bedrock.resource.models.entity.modelentity.geometry.bones.cubes.uv.South;
 import org.geysermc.pack.bedrock.resource.models.entity.modelentity.geometry.bones.cubes.uv.Up;
 import org.geysermc.pack.bedrock.resource.models.entity.modelentity.geometry.bones.cubes.uv.West;
-import org.geysermc.pack.converter.PackConversionContext;
-import org.geysermc.pack.converter.converter.Converter;
-import org.geysermc.pack.converter.data.ModelConversionData;
-import org.jetbrains.annotations.NotNull;
+import org.geysermc.pack.converter.converter.AssetCollector;
+import org.geysermc.pack.converter.converter.AssetExtractor;
+import org.geysermc.pack.converter.converter.CollectionContext;
+import org.geysermc.pack.converter.converter.ConversionContext;
+import org.geysermc.pack.converter.converter.ExtractionContext;
+import org.geysermc.pack.converter.converter.KeyedAssetConverter;
 import team.unnamed.creative.ResourcePack;
 import team.unnamed.creative.base.CubeFace;
 import team.unnamed.creative.model.Element;
@@ -58,136 +59,152 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
-@AutoService(Converter.class)
-public class ModelConverter implements Converter<ModelConversionData> {
+public class ModelConverter implements AssetExtractor<Model>, KeyedAssetConverter<Model, BedrockModel>, AssetCollector<BedrockModel> {
+    public static final ModelConverter INSTANCE = new ModelConverter();
+
     private static final String FORMAT_VERSION = "1.16.0";
     private static final String GEOMETRY_FORMAT = "geometry.%s";
 
     private static final float[] ELEMENT_OFFSET = new float[] { 8, 0, 8 };
 
     @Override
-    public void convert(@NotNull PackConversionContext<ModelConversionData> context) throws Exception {
-        ResourcePack javaPack = context.javaResourcePack();
-        BedrockResourcePack bedrockPack = context.bedrockResourcePack();
-        Collection<Model> models = javaPack.models();
-        if (models.isEmpty()) {
-            return;
+    public Collection<Model> extract(ResourcePack pack, ExtractionContext context) {
+        ModelStitcher.Provider modelProvider = context.vanillaPack()
+                .map(vanilla -> ModelStitcher.vanillaProvider(pack, vanilla))
+                .orElseGet(() -> ModelStitcher.baseProvider(pack));
+        // TODO maybe parallel, if model stitching takes a lot of time
+        return pack.models().stream()
+                .map(model -> new ModelStitcher(modelProvider, model, context.logListener()).stitch())
+                .toList();
+    }
+
+    @Override
+    public BedrockModel convert(Model model, ConversionContext context) throws Exception {
+        List<Element> elements = model.elements();
+        if (elements.isEmpty()) {
+            context.debug("Model " + model.key().key() + " has no elements");
+            return null;
         }
 
-        ModelStitcher.Provider provider = context.data().getModelProvider();
-        for (Model model : models) {
-            model = new ModelStitcher(provider, model, context.logListener()).stitch();
+        String value = model.key().value();
+        context.debug("Converting model " + model.key().key() + ":" + value);
 
-            List<Element> elements = model.elements();
-            if (elements.isEmpty()) {
-                context.debug("Model " + model.key().key() + " has no elements");
-                continue;
-            }
+        // TODO: Convert item models but save differently?
+        if (value.startsWith("item/")) {
+            return null;
+        }
 
-            String value = model.key().value();
-            context.debug("Converting model " + model.key().key() + ":" + value);
+        ModelEntity modelEntity = new ModelEntity();
+        modelEntity.formatVersion(FORMAT_VERSION);
 
-            // TODO: Convert item models but save differently?
-            if (value.startsWith("item/")) {
-                continue;
-            }
+        Geometry geometry = new Geometry();
 
-            ModelEntity modelEntity = new ModelEntity();
-            modelEntity.formatVersion(FORMAT_VERSION);
+        String namespace = model.key().namespace();
+        String fileName = value.substring(value.lastIndexOf('/') + 1);
 
-            Geometry geometry = new Geometry();
+        String geoName = (namespace.equals(Key.MINECRAFT_NAMESPACE) ? "" : namespace + ".") + fileName;
 
-            String namespace = model.key().namespace();
-            String fileName = value.substring(value.lastIndexOf('/') + 1);
+        // TODO: Don't hardcode all this
+        Description description = new Description();
+        description.identifier(String.format(GEOMETRY_FORMAT, geoName));
+        description.textureWidth(16);
+        description.textureHeight(16);
+        description.visibleBoundsWidth(2);
+        description.visibleBoundsHeight(2);
+        description.visibleBoundsOffset(new float[] { 0.0f, 0.25f, 0.0f });
+        geometry.description(description);
 
-            String geoName = (namespace.equals(Key.MINECRAFT_NAMESPACE) ? "" : namespace + ".") + fileName;
+        List<Bones> bones = new ArrayList<>();
 
-            // TODO: Don't hardcode all this
-            Description description = new Description();
-            description.identifier(String.format(GEOMETRY_FORMAT, geoName));
-            description.textureWidth(16);
-            description.textureHeight(16);
-            description.visibleBoundsWidth(2);
-            description.visibleBoundsHeight(2);
-            description.visibleBoundsOffset(new float[] { 0.0f, 0.25f, 0.0f });
-            geometry.description(description);
+        // TODO: Should each element be its own bone rather
+        //       than its own cube in the same bone?
+        int i = 0;
+        for (Element element : elements) {
+            float[] from = element.from().toArray();
+            float[] to = element.to().toArray();
 
-            List<Bones> bones = new ArrayList<>();
+            Bones bone = new Bones();
+            bone.name("bone_" + i++);
+            bone.pivot(new float[] { ELEMENT_OFFSET[0], ELEMENT_OFFSET[1], -ELEMENT_OFFSET[2] });
 
-            // TODO: Should each element be its own bone rather
-            //       than its own cube in the same bone?
-            int i = 0;
-            for (Element element : elements) {
-                float[] from = element.from().toArray();
-                float[] to = element.to().toArray();
+            Cubes cube = new Cubes();
+            cube.origin(new float[] { ELEMENT_OFFSET[0] - to[0], from[1], from[2] - ELEMENT_OFFSET[2] });
+            cube.size(new float[] { to[0] - from[0], to[1] - from[1], to[2] - from[2] });
 
-                Bones bone = new Bones();
-                bone.name("bone_" + i++);
-                bone.pivot(new float[] { ELEMENT_OFFSET[0], ELEMENT_OFFSET[1], -ELEMENT_OFFSET[2] });
+            ElementRotation elementRotation = element.rotation();
+            if (elementRotation != null) {
+                float[] origin = elementRotation.origin().toArray();
+                cube.pivot(new float[] { ELEMENT_OFFSET[0] - origin[0], ELEMENT_OFFSET[1] - origin[1], origin[2] - ELEMENT_OFFSET[2] });
 
-                Cubes cube = new Cubes();
-                cube.origin(new float[] { ELEMENT_OFFSET[0] - to[0], from[1], from[2] - ELEMENT_OFFSET[2] });
-                cube.size(new float[] { to[0] - from[0], to[1] - from[1], to[2] - from[2] });
-
-                ElementRotation elementRotation = element.rotation();
-                if (elementRotation != null) {
-                    float[] origin = elementRotation.origin().toArray();
-                    cube.pivot(new float[] { ELEMENT_OFFSET[0] - origin[0], ELEMENT_OFFSET[1] - origin[1], origin[2] - ELEMENT_OFFSET[2] });
-
-                    float angle = elementRotation.angle();
-                    float[] rotation = new float[3];
-                    switch (elementRotation.axis()) {
-                        case X -> rotation[0] = -angle;
-                        case Y -> rotation[1] = -angle;
-                        case Z -> rotation[2] = -angle;
-                    }
-
-                    cube.rotation(rotation);
+                float angle = elementRotation.angle();
+                float[] rotation = new float[3];
+                switch (elementRotation.axis()) {
+                    case X -> rotation[0] = -angle;
+                    case Y -> rotation[1] = -angle;
+                    case Z -> rotation[2] = -angle;
                 }
 
-                Uv uv = new Uv();
-                for (Map.Entry<CubeFace, ElementFace> entry : element.faces().entrySet()) {
-                    CubeFace face = entry.getKey();
-                    ElementFace elementFace = entry.getValue();
-                    if (elementFace.uv0() == null) {
-                        continue;
-                    }
+                cube.rotation(rotation);
+            }
 
-                    // The Java pack lib we use does this weird thing where it
-                    // divides the UV by 16, so we need to multiply it by 16
-
-                    String texture = elementFace.texture().replace("#", "");
-                    applyUv(uv, face, texture, multiplyUv(elementFace.uv0(), 16f));
+            Uv uv = new Uv();
+            for (Map.Entry<CubeFace, ElementFace> entry : element.faces().entrySet()) {
+                CubeFace face = entry.getKey();
+                ElementFace elementFace = entry.getValue();
+                if (elementFace.uv0() == null) {
+                    continue;
                 }
 
-                cube.uv(uv);
-                bone.cubes(List.of(cube));
+                // The Java pack lib we use does this weird thing where it
+                // divides the UV by 16, so we need to multiply it by 16
 
-                bones.add(bone);
+                String texture = elementFace.texture().replace("#", "");
+                applyUv(uv, face, texture, multiplyUv(elementFace.uv0(), 16f));
             }
 
-            geometry.bones(bones);
+            cube.uv(uv);
+            bone.cubes(List.of(cube));
 
-            modelEntity.geometry(List.of(geometry));
+            bones.add(bone);
+        }
 
-            if (model.key().namespace().contains("entity")) {
-                bedrockPack.addEntityModel(modelEntity, fileName + ".json");
-            } else {
-                // Bedrock only has a concept of entity or block models
-                bedrockPack.addBlockModel(modelEntity, fileName + ".json");
-            }
+        geometry.bones(bones);
 
-            context.data().addStitchedModel(model);
+        modelEntity.geometry(List.of(geometry));
+
+        if (model.key().namespace().contains("entity")) {
+            return new BedrockModel(BedrockModel.ModelType.ENTITY, fileName + ".json", modelEntity);
+        } else {
+            // Bedrock only has a concept of entity or block models
+            return new BedrockModel(BedrockModel.ModelType.BLOCK, fileName + ".json", modelEntity);
         }
     }
 
     @Override
-    public ModelConversionData createConversionData(@NotNull ConversionDataCreationContext context) {
-        return new ModelConversionData(
-            context.inputDirectory(), context.outputDirectory(),
-            ModelStitcher.vanillaProvider(context.javaResourcePack(), context.vanillaResourcePack()),
-            context.vanillaResourcePack()
-        );
+    public void include(BedrockResourcePack pack, List<BedrockModel> bedrockModels, CollectionContext context) {
+        List<String> entityModels = new ArrayList<>();
+        List<String> blockModels = new ArrayList<>();
+
+        for (BedrockModel model : bedrockModels) {
+            switch (model.type()) {
+                case ENTITY -> {
+                    if (entityModels.contains(model.fileName())) {
+                        context.warn("Conflicting entity model " + model.fileName() + "!");
+                        continue;
+                    }
+                    entityModels.add(model.fileName());
+                    pack.addEntityModel(model.model(), model.fileName());
+                }
+                case BLOCK -> {
+                    if (blockModels.contains(model.fileName())) {
+                        context.warn("Conflicting entity model " + model.fileName() + "!");
+                        continue;
+                    }
+                    blockModels.add(model.fileName());
+                    pack.addBlockModel(model.model(), model.fileName());
+                }
+            }
+        }
     }
 
     private TextureUV multiplyUv(TextureUV textureUV, float mult) {
