+++ converter/src/main/java/org/geysermc/pack/converter/converter/sound/SoundRegistryConverter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025 GeyserMC. http://geysermc.org
+ * Copyright (c) 2025-2025 GeyserMC. http://geysermc.org
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  *  of this software and associated documentation files (the "Software"), to deal
@@ -24,15 +24,15 @@
  *
  */
 
-package org.geysermc.pack.converter.newconverter.sound;
+package org.geysermc.pack.converter.converter.sound;
 
 import org.geysermc.pack.bedrock.resource.BedrockResourcePack;
 import org.geysermc.pack.bedrock.resource.sounds.sounddefinitions.SoundDefinitions;
 import org.geysermc.pack.bedrock.resource.sounds.sounddefinitions.Sounds;
-import org.geysermc.pack.converter.newconverter.AssetCollector;
-import org.geysermc.pack.converter.newconverter.CollectionContext;
-import org.geysermc.pack.converter.newconverter.ConversionContext;
-import org.geysermc.pack.converter.newconverter.NamespacedAssetConverter;
+import org.geysermc.pack.converter.converter.AssetCollector;
+import org.geysermc.pack.converter.converter.CollectionContext;
+import org.geysermc.pack.converter.converter.ConversionContext;
+import org.geysermc.pack.converter.converter.NamespacedAssetConverter;
 import org.jetbrains.annotations.Nullable;
 import team.unnamed.creative.sound.SoundEntry;
 import team.unnamed.creative.sound.SoundEvent;
@@ -43,8 +43,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class SoundRegistryConverter_ implements NamespacedAssetConverter<SoundRegistry, Map<String, SoundDefinitions>>, AssetCollector<Map<String, SoundDefinitions>> {
-    public static final SoundRegistryConverter_ INSTANCE = new SoundRegistryConverter_();
+public class SoundRegistryConverter implements NamespacedAssetConverter<SoundRegistry, Map<String, SoundDefinitions>>, AssetCollector<Map<String, SoundDefinitions>> {
+    public static final SoundRegistryConverter INSTANCE = new SoundRegistryConverter();
     static final String BEDROCK_SOUNDS_LOCATION = "sounds";
 
     @Override
