+++ converter/src/main/java/org/geysermc/pack/converter/converter/texture/transformer/TransformContext.java
@@ -29,7 +29,7 @@ package org.geysermc.pack.converter.converter.texture.transformer;
 import net.kyori.adventure.key.Key;
 import org.geysermc.pack.bedrock.resource.BedrockResourcePack;
 import org.geysermc.pack.converter.converter.texture.TextureMappings;
-import org.geysermc.pack.converter.newconverter.LogHelpers;
+import org.geysermc.pack.converter.converter.LogHelpers;
 import org.geysermc.pack.converter.util.ImageUtil;
 import org.geysermc.pack.converter.util.LogListener;
 import org.jetbrains.annotations.NotNull;
